Filter out elements to highlights, fix spacing.
This commit will filter out unecessary highlighted elements that caused overlap. It will also focus on spacing, reducing the amount of overlap on the currently highlighted elements.

The spacing itself *is* subjective to my own theme, and will most likely not 100% apply to all themes. However, this is in testing due to the behavior of spacing not being applied.

This commit also sets up the blueprints for ComboJudge.